requirements.txt
streamlit
numpy
tensorflow
opencv-python
fpdf
Pillow
reportlab
